+++ src/ha_bridge.cpp
@@ -110,6 +110,25 @@ String discoveryPayloadBinarySensor(const SettingsBundle& settings, const char* 
     return out;
 }
 
+String discoveryPayloadSwitch(const SettingsBundle& settings, const char* objectId, const char* name, const char* stateTopic, const char* commandTopicValue, const char* valueTemplate, const char* icon, const String& configurationUrl, const char* payloadOn, const char* payloadOff) {
+    JsonDocument doc;
+    doc["name"] = name;
+    doc["uniq_id"] = entityUniqueId(settings, objectId);
+    doc["stat_t"] = stateTopic;
+    doc["cmd_t"] = commandTopicValue;
+    doc["avty_t"] = availabilityTopic(settings);
+    doc["val_tpl"] = valueTemplate;
+    doc["pl_on"] = payloadOn;
+    doc["pl_off"] = payloadOff;
+    doc["stat_on"] = payloadOn;
+    doc["stat_off"] = payloadOff;
+    if (icon != nullptr) doc["ic"] = icon;
+    fillDevice(settings, doc["dev"].to<JsonObject>(), configurationUrl);
+    String out;
+    serializeJson(doc, out);
+    return out;
+}
+
 String discoveryPayloadNumber(const SettingsBundle& settings, const char* objectId, const char* name, const char* stateTopic, const char* commandTopicValue, int minValue, int maxValue, int step, const char* unit, const char* icon, const String& configurationUrl) {
     JsonDocument doc;
     doc["name"] = name;